refactor: StarRatingView -> KuStarRatingView 공통으로 사용할 수 있도록 수정
